Impliment System.CommandLine instead of manually parsing commands
