+++ Records.cs
@@ -1,0 +1,5 @@
+﻿namespace TimesheetTimothy;
+
+internal record Entry(string? JobCode, string? Hours, string? WorkType, string? Comments);
+internal record Day(Entry[]? Entries);
+internal record Week(Day? Monday, Day? Tuesday, Day? Wednesday, Day? Thursday, Day? Friday);
