+++ TimesheetTimothy.csproj
@@ -10,6 +10,7 @@
     <ItemGroup>
       <PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
       <PackageReference Include="Selenium.WebDriver" Version="4.1.0" />
+      <PackageReference Include="System.CommandLine" Version="2.0.0-beta3.22114.1" />
       <PackageReference Include="WebDriverManager" Version="2.12.3" />
     </ItemGroup>
